+++ libtelemetry/src/test/java/com/mapbox/android/telemetry/MapEventFactoryTest.java
@@ -142,9 +142,9 @@ public class MapEventFactoryTest {
   }
 
   private MapState obtainAValidMapState() {
-    float aLatitude = 40.416775f;
-    float aLongitude = -3.703790f;
-    float aZoom = 1.5f;
+    double aLatitude = 40.416775d;
+    double aLongitude = -3.703790d;
+    double aZoom = 1.5d;
     return new MapState(aLatitude, aLongitude, aZoom);
   }
 }
